Create LED Light Cycle
Trying to make a timer for a job with frequency. The person working needs to have a good visual of keeping time without a clock or timer. Idea is for LED light to light green for 20 seconds. Change to Yellow for 20, Then end with Red to show near end of completion. I have a buzzer activated once the time limit is reached. Trying to add a feature to use a button or switch or interrupt or stop the program from running. That is where I am stuck at.